% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countrylevel.R
\docType{data}
\name{countrylevel}
\alias{countrylevel}
\title{countrylevel: What a Waste Country level}
\format{
A tibble with 217 rows and 5 variables
\describe{
\item{iso3c}{NA}
\item{region_id}{NA}
\item{country_name}{NA}
\item{income_id}{NA}
\item{gdp}{NA}
}
}
\usage{
countrylevel
}
\description{
What a Waste is a global project to aggregate data on solid waste management
}
\keyword{datasets}
